+++ components/utils/unitTypeToIcon.ts
@@ -1,21 +1,19 @@
 import { type SelectableUnit } from 'appdeptus/models'
 import {
+  Anvil,
   Bus,
-  Car,
   Shield,
   UserRound,
   UsersRound,
-  VenetianMask,
   type LucideIcon
 } from 'lucide-react-native'
 
 const unitTypeToIcon = {
   character: UserRound,
   leader: Shield,
-  monster: VenetianMask,
   squad: UsersRound,
   transport: Bus,
-  vehicle: Car
+  vehicle: Anvil
 } as const satisfies Record<SelectableUnit['type'], LucideIcon>
 
 export default unitTypeToIcon