+++ modules/armies/screens/UnitSelection/UnitList.tsx
@@ -1,4 +1,4 @@
-import { Error, Loading, themeColors, VStack } from 'appdeptus/components'
+import { VStack } from 'appdeptus/components'
 import {
   type ArmyBuilder,
   type SelectableUnit,
@@ -7,40 +7,18 @@ import {
 import * as Crypto from 'expo-crypto'
 import React, { memo, useCallback, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
-import { FlatList, type ListRenderItem, RefreshControl } from 'react-native'
-import { useGetUnitListQuery } from '../../api'
+import { type ListRenderItem } from 'react-native'
+import { FlatList } from 'react-native-gesture-handler'
 import UnitCustomizationBottomSheet from './UnitCustomizationBottomSheet'
 import UnitListItem from './UnitListItem'
 import ref from './ref'
 
 type UnitListProps = {
-  type: 'character' | 'squad' | 'vehicle' | 'monster'
+  units: SelectableUnit[]
 }
 
-const UnitList = ({ type }: UnitListProps) => {
+const UnitList = ({ units }: UnitListProps) => {
   const { setValue, watch } = useFormContext<ArmyBuilder>()
-
-  const selectedCodex = watch('codex')
-
-  const { filteredData, isError, isFetching, refetch } = useGetUnitListQuery(
-    selectedCodex,
-    {
-      selectFromResult: ({ data, ...rest }) => ({
-        ...rest,
-        filteredData: data?.filter(({ type: t }) => {
-          switch (type) {
-            case 'character':
-              return t === 'character' || t === 'leader'
-            case 'monster':
-            case 'vehicle':
-              return t === 'monster' || t === 'vehicle' || t === 'transport'
-            default:
-              return t === type
-          }
-        })
-      })
-    }
-  )
 
   const handleAdd = useCallback(
     (unit: SelectableUnit) => {
@@ -55,7 +33,6 @@ const UnitList = ({ type }: UnitListProps) => {
         ...selectedUnits,
         {
           ...rest,
-          type,
           tier: unit.tiers[0],
           upgrades: [],
           selectionId: Crypto.randomUUID()
@@ -64,7 +41,7 @@ const UnitList = ({ type }: UnitListProps) => {
 
       setValue('points', points + unit.tiers[0].points)
     },
-    [setValue, type, watch]
+    [setValue, watch]
   )
 
   const [unitToEdit, setUnitToEdit] = useState<SelectableUnit | undefined>()
@@ -95,23 +72,11 @@ const UnitList = ({ type }: UnitListProps) => {
   return (
     <>
       <FlatList
-        data={filteredData}
-        contentContainerStyle={!filteredData?.length ? { flex: 1 } : undefined}
+        data={units}
+        contentContainerStyle={!units?.length ? { flex: 1 } : undefined}
         ItemSeparatorComponent={() => <VStack className='h-4' />}
-        ListEmptyComponent={() =>
-          isError ? <Error /> : isFetching ? <Loading /> : null
-        }
         ListFooterComponent={() => <VStack className='h-8' />}
         keyExtractor={({ id }) => String(id)}
-        refreshControl={
-          isError ? (
-            <RefreshControl
-              tintColor={themeColors.default.primary[300]}
-              refreshing={isFetching}
-              onRefresh={refetch}
-            />
-          ) : undefined
-        }
         renderItem={renderItem}
         showsVerticalScrollIndicator={false}
       />